:fire: Remove qa folder Move network test under test folder
