--- qa/.eslintignore
@@ -1,10 +1,0 @@
-docs/
-lisk/
-templates/
-types/
-examples/
-**/*.d.ts
-jest.config.js
-coverage
-benchmark
-.eslintrc.js
