--- qa/.gitignore
@@ -1,11 +1,0 @@
-# Dependency directory
-node_modules/
-
-# IDE
-.vscode/
-
-# Reports and logs
-output/
-
-#others
-.DS_Store
