--- qa/.lintstagedrc.json
@@ -1,5 +1,0 @@
-{
-	"*.js": ["prettier --write"],
-	"*.ts": ["prettier --write", "eslint"],
-	"*.{json,md}": ["prettier --write"]
-}
