--- qa/.nvmrc
@@ -1,1 +1,0 @@
-12.18.3
