--- qa/.prettierrc.json
@@ -1,7 +1,0 @@
-{
-	"printWidth": 100,
-	"singleQuote": true,
-	"trailingComma": "all",
-	"useTabs": true,
-	"arrowParens": "avoid"
-}
