--- qa/Jenkinsfile
@@ -1,138 +1,0 @@
-@Library('lisk-jenkins') _
-
-properties([
-  parameters([
-    string(name: 'SDK_BRANCH_NAME', defaultValue: 'development', description: 'Lisk core branch name', ),
-    string(name: 'CORE_BRANCH_NAME', defaultValue: 'development', description: 'Lisk core branch name', ),
-    string(name: 'NETWORK', defaultValue: 'jenkinsnet-${BUILD_ID}', description: 'To Run test against a network', ),
-    string(name: 'NODES_PER_REGION', defaultValue: '1', description: 'Number of nodes per region', ),
-    string(name: 'STRESS_COUNT', defaultValue: '500', description: 'Number of transactions to create', ), // Used by stage: Test Network Stress
-    booleanParam(name: 'CLEAN_UP', defaultValue: true, description: 'Destroy deployed nodes when done testing', ),
-  ])
-])
-
-pipeline {
-  agent { node { label 'lisk-core' } }
-  options { disableConcurrentBuilds() }
-  stages {
-    stage('Build') {
-      steps {
-        dir('qa') {
-          nvm(getNodejsVersion()) {
-            sh 'npm ci'
-          }
-        }
-      }
-    }
-    stage('Trigger core build') {
-      steps {
-        script {
-          def b = build job: 'lisk-qa/lisk-core-build-dev',
-                  parameters: [string(name: 'COMMITISH_SDK', value: """${params.SDK_BRANCH_NAME}"""),
-                   string(name: 'COMMITISH_CORE', value: """${params.CORE_BRANCH_NAME}""")]
-          env.LISK_VERSION = b.getBuildVariables().get('LISK_VERSION')
-        }
-      }
-    }
-    stage('Deploy network') {
-      steps {
-        retry(5) {
-          ansiColor('xterm') {
-            ansibleTower \
-              towerServer: 'tower',
-              templateType: 'job',
-              jobTemplate: '14',  // devnet-deploy
-              jobType: 'run',
-              extraVars: """devnet: ${params.NETWORK}
-do_nodes_per_region: ${params.NODES_PER_REGION}
-jenkins_ci: 'yes'
-lisk_version: ${env.LISK_VERSION}""",
-              importTowerLogs: true,
-              throwExceptionWhenFail: true,
-              verbose: false
-          }
-        }
-      }
-    }
-    stage('Generate peer config and enable forging') {
-      steps {
-        dir('qa') {
-          nvm(getNodejsVersion()) {
-            sh '''
-            npm run tools:peers:seed:node
-            npm run tools:peers:connected
-            npm run tools:peers:network:nodes
-            npm run tools:delegates:enable
-            '''
-          }
-        }
-      }
-    }
-    stage('Test Scenarios') {
-      steps {
-        retry(2) {
-          timestamps {
-            dir('qa') {
-              nvm(getNodejsVersion()) {
-                ansiColor('xterm') {
-                  sh 'npm run features:debug || true'
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    stage('Test Network Stress') {
-      steps {
-        timestamps {
-          dir('qa') {
-            nvm(getNodejsVersion()) {
-              ansiColor('xterm') {
-                sh 'npm run stress:generic || true'
-                sh 'npm run stress:diversified'
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-  post {
-    always {
-      allure includeProperties: false, jdk: '', results: [[path: 'qa/output']]
-    }
-    failure {
-      script {
-        if (env.CLEAN_UP == "true") {
-          ansibleTower \
-            towerServer: 'tower',
-            templateType: 'job',
-            jobTemplate: '16',  // devnet-archive-logs
-            jobType: 'run',
-            extraVars: """devnet: ${params.NETWORK}""",
-            throwExceptionWhenFail: false,
-            verbose: false
-        }
-      }
-      liskSlackSend('danger', "Build failed (<${env.BUILD_URL}/console|console>, <${env.BUILD_URL}/changes|changes>)", 'lisk-ci-core')
-    }
-    cleanup {
-      script {
-        if (env.CLEAN_UP == "true") {
-          ansibleTower \
-            towerServer: 'tower',
-            templateType: 'job',
-            jobTemplate: '15',  // do-destroy-tag
-            jobType: 'run',
-            extraVars: """do_tag: ${params.NETWORK}_node""",
-            throwExceptionWhenFail: false,
-            verbose: false
-        } else {
-          echo 'Not cleaning up, as requested.'
-        }
-      }
-    }
-  }
-}
-// vim: filetype=groovy
