--- qa/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "lisk-core-qa",
-  "version": "0.1.0",
-  "description": "Lisk core quality assurance",
-  "author": "Lisk Foundation <admin@lisk.io>, lightcurve GmbH <admin@lightcurve.io>",
-  "license": "Apache-2.0",
-  "keywords": [
-    "qa",
-    "lisk",
-    "blockchain"
-  ],
-  "homepage": "https://github.com/LiskHQ/lisk-core/blob/development/qa/README.md",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/LiskHQ/lisk-core.git"
-  },
-  "bugs": {
-    "url": "https://github.com/LiskHQ/lisk-core/issues"
-  },
-  "engines": {
-    "node": ">=12.18.3",
-    "npm": ">=6.13.4"
-  },
-  "scripts": {
-    "lint": "eslint --ext .js,.ts .",
-		"lint:fix": "eslint --fix --ext .js,.ts .",
-		"format": "prettier --write '**/*'",
-		"prepack": "rm -rf lib && tsc -b && oclif-dev manifest && oclif-dev readme",
-		"prebuild": "if test -d dist; then rm -r dist; fi;",
-		"build": "tsc"
-  },
-  "dependencies": {
-    "lisk-sdk": "5.0.0-debug.0"
-  },
-  "devDependencies": {
-    "@types/node": "12.12.11",
-    "@typescript-eslint/eslint-plugin": "3.5.0",
-    "@typescript-eslint/parser": "3.5.0",
-    "ansi-regex": "5.0.0",
-    "eslint": "7.3.1",
-    "eslint-config-airbnb-base": "14.2.0",
-    "eslint-config-lisk-base": "1.2.2",
-    "eslint-config-prettier": "6.11.0",
-    "eslint-plugin-chai-expect": "2.1.0",
-    "eslint-plugin-import": "2.22.0",
-    "eslint-plugin-jest": "23.17.1",
-    "eslint-plugin-mocha": "7.0.1",
-    "husky": "4.2.5",
-    "istanbul-lib-coverage": "3.0.0",
-    "istanbul-lib-report": "3.0.0",
-    "istanbul-reports": "3.0.0",
-    "lerna": "3.22.1",
-    "lint-staged": "10.2.11",
-    "prettier": "2.0.5",
-    "typescript": "3.8.3",
-    "yarn": "1.22.4"
-  }
-}
