--- qa/scripts/run.sh
@@ -1,5 +1,0 @@
-#! /bin/bash
-
-npm run tools:peers:seed:node
-npm run features || true
-npm run stress:generic
