--- qa/tsconfig.json
@@ -1,25 +1,0 @@
-{
-	"compilerOptions": {
-		"forceConsistentCasingInFileNames": true,
-		"target": "es2019",
-		"module": "commonjs",
-		"moduleResolution": "node",
-		"newLine": "lf",
-		"importHelpers": true,
-		"noFallthroughCasesInSwitch": true,
-		"noImplicitReturns": true,
-		"noUnusedLocals": true,
-		"noUnusedParameters": true,
-		"pretty": true,
-		"removeComments": true,
-		"resolveJsonModule": true,
-		"sourceMap": true,
-		"strict": true,
-		"composite": true,
-		"declaration": true,
-		"noImplicitAny": false,
-		"rootDir": "./src",
-		"outDir": "./dist"
-	},
-	"include": ["./src/**/*.ts", "./src/**/*.json"]
-}
